Task 20: update JobSeekerApplyController for searching and applying for job

+++ src/main/java/com/job/portal/web/application/payload/IRecruiterJobs.java
@@ -11,7 +11,7 @@ public interface IRecruiterJobs {
 
     Long getTotalCandidates();
 
-    int getJob_post_id();
+    long getJob_post_id();
 
     String getJob_title();
